refactor: settlement의 cotroller, dto 내부 패키지 분리
