+++ src/main/java/com/zero/cohousesever/settlement/controller/SettlementController.java
@@ -5,10 +5,10 @@ import com.zero.cohousesever.common.exception.ErrorCode;
 import com.zero.cohousesever.file.dto.FileUploadResponse;
 import com.zero.cohousesever.file.service.S3Service;
 import com.zero.cohousesever.member.security.CustomUserDetails;
-import com.zero.cohousesever.settlement.dto.CreateSettlementRequest;
-import com.zero.cohousesever.settlement.dto.ParticipantResponse;
-import com.zero.cohousesever.settlement.dto.SettlementHistoryResponse;
-import com.zero.cohousesever.settlement.dto.SettlementResponse;
+import com.zero.cohousesever.settlement.dto.settlement.CreateSettlementRequest;
+import com.zero.cohousesever.settlement.dto.payment.ParticipantResponse;
+import com.zero.cohousesever.settlement.dto.settlement.SettlementHistoryResponse;
+import com.zero.cohousesever.settlement.dto.settlement.SettlementResponse;
 import com.zero.cohousesever.settlement.service.PaymentService;
 import com.zero.cohousesever.settlement.service.SettlementService;
 import lombok.RequiredArgsConstructor;
