+++ src/main/java/com/zero/cohousesever/settlement/dto/settlement/OcrResult.java
@@ -1,4 +1,4 @@
-package com.zero.cohousesever.settlement.dto;
+package com.zero.cohousesever.settlement.dto.settlement;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
