+++ src/main/java/com/zero/cohousesever/settlement/dto/settlement/SettlementResponse.java
@@ -1,8 +1,9 @@
-package com.zero.cohousesever.settlement.dto;
+package com.zero.cohousesever.settlement.dto.settlement;
 
-import com.zero.cohousesever.settlement.entity.Settlement;
-import com.zero.cohousesever.settlement.entity.SettlementCategory;
-import com.zero.cohousesever.settlement.entity.SettlementStatus;
+import com.zero.cohousesever.settlement.dto.payment.ParticipantResponse;
+import com.zero.cohousesever.settlement.entity.settlement.Settlement;
+import com.zero.cohousesever.settlement.entity.settlement.SettlementCategory;
+import com.zero.cohousesever.settlement.entity.settlement.SettlementStatus;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
