+++ src/main/java/com/zero/cohousesever/settlement/repository/SettlementParticipantRepository.java
@@ -1,8 +1,8 @@
 package com.zero.cohousesever.settlement.repository;
 
 import com.zero.cohousesever.member.entity.Member;
-import com.zero.cohousesever.settlement.entity.Settlement;
-import com.zero.cohousesever.settlement.entity.SettlementParticipant;
+import com.zero.cohousesever.settlement.entity.settlement.Settlement;
+import com.zero.cohousesever.settlement.entity.settlement.SettlementParticipant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
