+++ src/main/java/com/zero/cohousesever/settlement/repository/SettlementRepository.java
@@ -2,8 +2,8 @@ package com.zero.cohousesever.settlement.repository;
 
 import com.zero.cohousesever.group.entity.Group;
 import com.zero.cohousesever.member.entity.Member;
-import com.zero.cohousesever.settlement.entity.Settlement;
-import com.zero.cohousesever.settlement.entity.SettlementStatus;
+import com.zero.cohousesever.settlement.entity.settlement.Settlement;
+import com.zero.cohousesever.settlement.entity.settlement.SettlementStatus;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
