+++ src/main/java/com/zero/cohousesever/settlement/service/SettlementService.java
@@ -13,8 +13,17 @@ import com.zero.cohousesever.group.repository.GroupRepository;
 import com.zero.cohousesever.member.entity.Member;
 import com.zero.cohousesever.member.repository.MemberRepository;
 import com.zero.cohousesever.ocr.TesseractOcrService;
-import com.zero.cohousesever.settlement.dto.*;
-import com.zero.cohousesever.settlement.entity.*;
+import com.zero.cohousesever.settlement.dto.payment.ParticipantResponse;
+import com.zero.cohousesever.settlement.dto.settlement.CreateSettlementRequest;
+import com.zero.cohousesever.settlement.dto.settlement.OcrResult;
+import com.zero.cohousesever.settlement.dto.settlement.SettlementHistoryResponse;
+import com.zero.cohousesever.settlement.dto.settlement.SettlementResponse;
+import com.zero.cohousesever.settlement.entity.payment.PaymentHistory;
+import com.zero.cohousesever.settlement.entity.payment.PaymentStatus;
+import com.zero.cohousesever.settlement.entity.settlement.Settlement;
+import com.zero.cohousesever.settlement.entity.settlement.SettlementHistory;
+import com.zero.cohousesever.settlement.entity.settlement.SettlementParticipant;
+import com.zero.cohousesever.settlement.entity.settlement.SettlementStatus;
 import com.zero.cohousesever.settlement.event.SettlementCreatedEvent;
 import com.zero.cohousesever.settlement.repository.PaymentHistoryRepository;
 import com.zero.cohousesever.settlement.repository.SettlementHistoryRepository;
