+++ src/test/java/com/zero/cohousesever/settlement/service/SettlementServiceTest.java
@@ -9,9 +9,11 @@ import com.zero.cohousesever.group.repository.GroupMemberRepository;
 import com.zero.cohousesever.group.repository.GroupRepository;
 import com.zero.cohousesever.member.entity.Member;
 import com.zero.cohousesever.member.repository.MemberRepository;
-import com.zero.cohousesever.settlement.dto.CreateSettlementRequest;
-import com.zero.cohousesever.settlement.dto.SettlementResponse;
-import com.zero.cohousesever.settlement.entity.*;
+import com.zero.cohousesever.settlement.dto.settlement.CreateSettlementRequest;
+import com.zero.cohousesever.settlement.dto.settlement.SettlementResponse;
+import com.zero.cohousesever.settlement.entity.payment.PaymentHistory;
+import com.zero.cohousesever.settlement.entity.payment.PaymentStatus;
+import com.zero.cohousesever.settlement.entity.settlement.*;
 import com.zero.cohousesever.settlement.repository.PaymentHistoryRepository;
 import com.zero.cohousesever.settlement.repository.SettlementHistoryRepository;
 import com.zero.cohousesever.settlement.repository.SettlementParticipantRepository;
